+++ test/hidden-prs.test.js
@@ -55,6 +55,7 @@ function checkEngine(estadoInicial) {
   e.enrichMyPRBranches = async () => { };
   e.refreshMergeStates = async () => { };
   e.refreshStaleStates = async () => { };
+  e.refreshReviewSignals = async () => { }; // sinal por ref é rede: fora do ciclo de teste
   e.scanPushbacks = async () => { };
   e.checkUpdate = async () => { };
   e.schedule = () => { };