+++ test/retry-net.test.js
@@ -478,6 +478,7 @@ function checkEngineG15() {
   e.enrichMyPRBranches = async () => { };
   e.refreshMergeStates = async () => { };
   e.refreshStaleStates = async () => { };
+  e.refreshReviewSignals = async () => { }; // sinal por ref é rede: fora do ciclo de teste
   e.scanPushbacks = async () => { };
   e.checkUpdate = async () => { };
   e.launchReview = () => { throw new Error('launchReview não deveria rodar neste teste'); };